+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Removed `https://derpicdn.net/img/*` from Derpibooru's oEmbed URL scheme.
 - Text contents of `<mf-figure-caption>` are now wrapped with `<p>`.
 - Fixed a parsing issue that a closing parenthesis (`)`) of a media tag was sometimes overlooked.
+- The table of contents pane now supports keyboard navigation.
 
 ## [0.0.5] - 2018-12-30